# API документация/openapi.json
{"openapi":"3.1.0","info":{"title":"MedPartners API","version":"1.0.0"},"paths":{"/dashboard/stats":{"get":{"summary":"Статистика для дашборда","operationId":"dashboard_stats_dashboard_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardStats"}}}}}}},"/upload-archive":{"post":{"summary":"Загрузить ZIP-архив с прайсами","operationId":"upload_archive_upload_archive_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_archive_upload_archive_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/documents/{doc_id}/status":{"get":{"summary":"Статус обработки документа","operationId":"get_status_documents__doc_id__status_get","parameters":[{"name":"doc_id","in":"path","required":true,"schema":{"type":"string","title":"Doc Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentStatusResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/services":{"get":{"summary":"Список услуг справочника","operationId":"list_services_services_get","parameters":[{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ServiceOut"},"title":"Response List Services Services Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Создать услугу в справочнике","operationId":"create_service_services_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/services/{service_id}/partners":{"get":{"summary":"Партнёры, оказывающие услугу, с ценами","operationId":"service_partners_services__service_id__partners_get","parameters":[{"name":"service_id","in":"path","required":true,"schema":{"type":"string","title":"Service Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PriceItemOut"},"title":"Response Service Partners Services  Service Id  Partners Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/catalog/upload":{"post":{"summary":"Массовая загрузка справочника услуг (JSON)","operationId":"upload_catalog_catalog_upload_post","requestBody":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CatalogEntry"},"type":"array","title":"Entries"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CatalogUploadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/partners":{"get":{"summary":"Список партнёров","operationId":"list_partners_partners_get","parameters":[{"name":"city","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"}},{"name":"is_active","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PartnerOut"},"title":"Response List Partners Partners Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/partners/{partner_id}/services":{"get":{"summary":"Все услуги партнёра с ценами","operationId":"partner_services_partners__partner_id__services_get","parameters":[{"name":"partner_id","in":"path","required":true,"schema":{"type":"string","title":"Partner Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PriceItemOut"},"title":"Response Partner Services Partners  Partner Id  Services Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search":{"get":{"summary":"Полнотекстовый поиск","operationId":"search_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","title":"Q"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/unmatched":{"get":{"summary":"Несопоставленные позиции прайса","operationId":"get_unmatched_unmatched_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PriceItemOut"},"title":"Response Get Unmatched Unmatched Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/match":{"post":{"summary":"Ручное сопоставление позиции с услугой","operationId":"match_item_match_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatchRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PriceItemOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_archive_upload_archive_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_archive_upload_archive_post"},"CatalogEntry":{"properties":{"service_name":{"type":"string","title":"Service Name"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"icd_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Icd Code"},"synonyms":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Synonyms"}},"type":"object","required":["service_name"],"title":"CatalogEntry"},"CatalogUploadResponse":{"properties":{"created":{"type":"integer","title":"Created"},"updated":{"type":"integer","title":"Updated"}},"type":"object","required":["created","updated"],"title":"CatalogUploadResponse"},"DashboardStats":{"properties":{"total_docs":{"type":"integer","title":"Total Docs"},"done_docs":{"type":"integer","title":"Done Docs"},"error_docs":{"type":"integer","title":"Error Docs"},"needs_review_docs":{"type":"integer","title":"Needs Review Docs"},"total_items":{"type":"integer","title":"Total Items"},"unmatched_items":{"type":"integer","title":"Unmatched Items"},"verified_items":{"type":"integer","title":"Verified Items"},"total_partners":{"type":"integer","title":"Total Partners"},"total_services":{"type":"integer","title":"Total Services"},"normalization_pct":{"type":"number","title":"Normalization Pct"}},"type":"object","required":["total_docs","done_docs","error_docs","needs_review_docs","total_items","unmatched_items","verified_items","total_partners","total_services","normalization_pct"],"title":"DashboardStats"},"DocumentStatusResponse":{"properties":{"status":{"type":"string","title":"Status"},"items_extracted":{"type":"integer","title":"Items Extracted"},"log":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Log"}},"type":"object","required":["status","items_extracted","log"],"title":"DocumentStatusResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MatchRequest":{"properties":{"item_id":{"type":"string","title":"Item Id"},"service_id":{"type":"string","title":"Service Id"},"note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Note"}},"type":"object","required":["item_id","service_id"],"title":"MatchRequest"},"PartnerOut":{"properties":{"partner_id":{"type":"string","format":"uuid","title":"Partner Id"},"name":{"type":"string","title":"Name"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"contact_email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Contact Email"},"contact_phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Contact Phone"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["partner_id","name","city","address","contact_email","contact_phone","is_active"],"title":"PartnerOut"},"PriceItemOut":{"properties":{"item_id":{"type":"string","format":"uuid","title":"Item Id"},"service_name_raw":{"type":"string","title":"Service Name Raw"},"service_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Service Id"},"price_original":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Price Original"},"price_resident_kzt":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Price Resident Kzt"},"price_nonresident_kzt":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Price Nonresident Kzt"},"currency_original":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Currency Original"},"is_verified":{"type":"boolean","title":"Is Verified"},"verification_note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Verification Note"},"effective_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Effective Date"},"partner_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Partner Id"},"doc_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Doc Id"},"price_anomaly":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Price Anomaly","default":false}},"type":"object","required":["item_id","service_name_raw","service_id","price_original","price_resident_kzt","price_nonresident_kzt","currency_original","is_verified","verification_note","effective_date","partner_id","doc_id"],"title":"PriceItemOut"},"SearchResponse":{"properties":{"services":{"items":{"$ref":"#/components/schemas/ServiceOut"},"type":"array","title":"Services"},"partners":{"items":{"$ref":"#/components/schemas/PartnerOut"},"type":"array","title":"Partners"},"price_items":{"items":{"$ref":"#/components/schemas/PriceItemOut"},"type":"array","title":"Price Items"}},"type":"object","required":["services","partners","price_items"],"title":"SearchResponse"},"ServiceCreate":{"properties":{"service_name":{"type":"string","title":"Service Name"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"icd_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Icd Code"},"synonyms":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Synonyms"}},"type":"object","required":["service_name"],"title":"ServiceCreate"},"ServiceOut":{"properties":{"service_id":{"type":"string","format":"uuid","title":"Service Id"},"service_name":{"type":"string","title":"Service Name"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"icd_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Icd Code"},"synonyms":{"anyOf":[{},{"type":"null"}],"title":"Synonyms"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["service_id","service_name","category","icd_code","synonyms","is_active"],"title":"ServiceOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}